add new concurrent base learner
