+++ orangecontrib/recommendation/widgets/owtrustsvd.py
@@ -1,39 +1,17 @@
-import copy
-from types import SimpleNamespace as namespace
-
 from AnyQt.QtCore import Qt
 
 from Orange.data import Table
 from Orange.widgets import settings, gui
 from Orange.widgets.utils.signals import Output, Input
 from Orange.widgets.utils.widgetpreview import WidgetPreview
-from Orange.widgets.utils.owlearnerwidget import OWBaseLearner
-from Orange.widgets.utils.concurrent import TaskState, ConcurrentWidgetMixin
 
 from orangecontrib.recommendation import TrustSVDLearner
 from orangecontrib.recommendation.utils import format_data
+from orangecontrib.recommendation.widgets.utils.owconcurrentlearner import OWConcurrentLearner
 import orangecontrib.recommendation.optimizers as opt
 
 
-class Result(namespace):
-    model = None
-
-
-def run(data, learner, max_iter, state: TaskState):
-    def callback(iter):
-        nonlocal max_iter
-        nonlocal state
-        state.set_progress_value(int(iter / max_iter * 100))
-        return state.is_interruption_requested()
-
-    _learner = copy.copy(learner)
-    _learner.callback = callback
-
-    model = _learner(data)
-    return Result(model=model)
-
-
-class OWTrustSVD(OWBaseLearner, ConcurrentWidgetMixin):
+class OWTrustSVD(OWConcurrentLearner):
     # Widget needs a name, or it is considered an abstract widget
     # and not shown in the menu.
     name = "TrustSVD"
@@ -46,10 +24,10 @@ class OWTrustSVD(OWBaseLearner, ConcurrentWidgetMixin):
 
     MISSING_DATA_WARNING = 0
 
-    class Inputs(OWBaseLearner.Inputs):
+    class Inputs(OWConcurrentLearner.Inputs):
         trust = Input("Trust information", Table)
 
-    class Outputs(OWBaseLearner.Outputs):
+    class Outputs(OWConcurrentLearner.Outputs):
         p = Output("P", Table, explicit=True)
         q = Output("Q", Table, explicit=True)
         y = Output("Y", Table, explicit=True)
@@ -81,10 +59,6 @@ class OWTrustSVD(OWBaseLearner, ConcurrentWidgetMixin):
     rho = settings.Setting(0.9)
     beta1 = settings.Setting(0.9)
     beta2 = settings.Setting(0.999)
-
-    def __init__(self, preprocessors=None):
-        ConcurrentWidgetMixin.__init__(self)
-        OWBaseLearner.__init__(self, preprocessors=preprocessors)
 
     def add_main_layout(self):
         # hbox = gui.hBox(self.controlArea, "Settings")
@@ -166,10 +140,6 @@ class OWTrustSVD(OWBaseLearner, ConcurrentWidgetMixin):
                                       label="Seed:", alignment=Qt.AlignRight,
                                       callback=self.settings_changed)
         self.settings_changed()  # Update (extra) settings
-
-    def setup_layout(self):
-        super().setup_layout()
-        gui.button(self.apply_button, self, "Cancel", callback=self.cancel)
 
     def settings_changed(self):
         # Enable/Disable Fixed seed control
@@ -298,27 +268,6 @@ class OWTrustSVD(OWBaseLearner, ConcurrentWidgetMixin):
         if self.valid_data or self.data is None:
             super().update_learner()
 
-    def update_model(self):
-        self.show_fitting_failed(None)
-        self.model = None
-        if self.check_data() and self._check_data():
-            self.start(run, self.data, self.learner, self.num_iter)
-
-    def cancel(self):
-        super().cancel()
-        self.model = None
-        self.commit()
-
-    def on_done(self, result: Result):
-        self.model = result.model
-        self.model.name = self.learner_name or self.name
-        self.model.instances = self.data
-        self.commit()
-
-    def on_exception(self, ex: Exception):
-        self.show_fitting_failed(ex)
-        self.commit()
-
     def commit(self):
         self.Outputs.model.send(self.model)
 
@@ -342,10 +291,6 @@ class OWTrustSVD(OWBaseLearner, ConcurrentWidgetMixin):
         self.trust = trust
         if self.auto_apply:
             self.apply()
-
-    def onDeleteWidget(self):
-        self.shutdown()
-        super().onDeleteWidget()
 
 
 if __name__ == '__main__':
